Break Search bar into its own component
